+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - If JsonArtemisSerializer#setUsePrototypes is true (default), omit components made
     up of default values.
   - Json DOM was parsed multiple times, fixed.
+  - Overloaded JsonArtemisSerializer#load - deserialize directly from JsonValue.
 
 #### Version: 1.2.1 - 2015-11-23
 - **Fix**: `@PooledWeaver` properly resets primitives and strings back to initial values.